# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
        "url": "/"
    }],
...
    "paths": {
...
        "/demarcheur/request/all": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Get all request",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": ""
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/demarcheur/all": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Get all demarcheur list",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": ""
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/demarcheur/request/answer": {
            "post": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Answer to demande",
                "requestBody": {
                    "description": "Answer to demande. The answer is 'accepted' or 'rejected'",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/request_answer"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/request_answer"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/demarcheur/appart/all": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Get all appart",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": ""
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/demarcheur/invitation/code": {
            "get": {
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Get demarcheur code ",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": ""
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/demarcheur/invitation/appart/all": {
            "get": {
                "parameters": [{
                    "name": "demarcheur_code",
                    "in": "query",
                    "required": false,
                    "description": "le code du demarcheur",
                    "type": "string"
                }],
                "security": [{
                    "bearerAuth": []
                }],
                "tags": [
                    "Demarcheur request"
                ],
                "summary": "Permet de voir recuperer les appartements gere par un demarcheur",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Cette erreur se produit quand il y'a un problème côté client"
                    },
                    "401": {
                        "description": "Cette erreur se produit quand l'utilisateur n'est pas autorisé"
                    },
                    "404": {
                        "description": "Cette erreur se produit quand la ressource demandée n'a pas été trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "request_answer": {
                "type": "object",
                "properties": {
                    "request_id": {
                        "type": "string",
                        "format": "string"
                    },
                    "answer": {
                        "type": "string",
                        "format": "string"
                    }
                }
            },
...
        }
    }
}